Fusión vista responsive, con lógica de login
