--- src/modelo/GestionProyecto.java
@@ -1,14 +1,0 @@
-package modelo;
-
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-
-/**
- *
- * @author josem
- */
-public class GestionProyecto {
-    
-}
