+++ src/vista/vistalogin.css
@@ -9,16 +9,16 @@
 .buttonDefault {
     -fx-text-fill : white;
     -fx-font-family : 'Poppins';
-    -fx-font-size : 22px;
+    -fx-font-size : 22 em;
     -fx-background-color : #f9560b;
-    -fx-background-radius : 12px;
+    -fx-background-radius : 15px;
     -fx-min-height : 3em;
 }
 
 .buttonDefault:pressed {
     -fx-text-fill : white;
     -fx-font-family : 'Poppins';
-    -fx-font-size : 22px;
+    -fx-font-size : 22 em;
     -fx-background-color : #ff8a54;
     -fx-background-radius : 12px;
 }
@@ -27,19 +27,19 @@
     -fx-text-fill : white;
     -fx-font-family : 'Poppins';
     -fx-font-weight : bold;
-    -fx-font-size : 35px;
+    -fx-font-size : 35 em;
 }
 
 .labelRegular {
     -fx-text-fill : white;
     -fx-font-family : 'Poppins';
-    -fx-font-size : 18px;
+    -fx-font-size : 18 em;
 }
 
 .textField {
     -fx-text-fill : black;
     -fx-font-family : 'Poppins';
-    -fx-font-size : 18px;
+    -fx-font-size : 18 em;
     -fx-background-radius : 12px;
     -fx-background-color: #ececec;
 }
@@ -47,7 +47,7 @@
 .textField:focused {
     -fx-text-fill : black;
     -fx-font-family : 'Poppins';
-    -fx-font-size : 18px;
+    -fx-font-size : 18 em;
     -fx-background-radius : 12px;
     -fx-background-color: white;
 }
@@ -55,7 +55,7 @@
 .passField {
     -fx-text-fill : black;
     -fx-font-family : 'Poppins';
-    -fx-font-size : 14px;
+    -fx-font-size : 14 em;
     -fx-background-radius : 12px;
     -fx-background-color: #ececec;
 }
@@ -63,7 +63,7 @@
 .passField:focused {
     -fx-text-fill : black;
     -fx-font-family : 'Poppins';
-    -fx-font-size : 14px;
+    -fx-font-size : 14 em;
     -fx-background-radius : 12px;
     -fx-background-color: white;
 }
@@ -71,7 +71,7 @@
 .checkBox {
     -fx-text-fill : white;
     -fx-font-family : 'Poppins';
-    -fx-font-size : 16px;
+    -fx-font-size : 16 em;
 }
 
 .checkBox .box {
